Factor out a testable module
